third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-time-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <time> with iterationComposite
  assert_equals: expected "250s" but got "50s"
Harness: the test ran to completion.
